third_party/WebKit/LayoutTests/fast/webgl/texImage-imageBitmap-from-image-resize-expected.txt

CONSOLE MESSAGE: line 228: x: 0, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 0, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 144, y: 0, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 160, y: 0, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 176, y: 0, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 192, y: 0, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 208, y: 0, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 224, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 0, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 16, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 32, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 32, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 336, y: 32, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 352, y: 32, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 368, y: 32, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 0, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 16, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 32, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 48, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 64, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 80, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 96, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 112, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 128, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 144, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 160, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 176, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 192, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 208, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 224, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 240, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 256, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 272, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 288, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 304, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 320, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 336, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 352, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 368, y: 48, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 0, y: 64, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 16, y: 64, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 32, y: 64, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 48, y: 64, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 64, y: 64, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 80, y: 64, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 96, y: 64, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 112, y: 64, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 128, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 64, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 80, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 96, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 80, y: 96, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 96, y: 96, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 112, y: 96, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 128, y: 96, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 144, y: 96, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 160, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 96, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 112, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 128, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 128, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 272, y: 128, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 288, y: 128, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 304, y: 128, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 320, y: 128, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 336, y: 128, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 352, y: 128, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 368, y: 128, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 0, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 16, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 32, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 48, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 64, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 80, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 96, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 112, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 128, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 144, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 160, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 176, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 192, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 208, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 224, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 240, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 256, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 272, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 288, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 304, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 320, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 336, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 352, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 368, y: 144, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 0, y: 160, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 16, y: 160, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 32, y: 160, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 48, y: 160, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 64, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 160, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 176, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 192, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 16, y: 192, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 32, y: 192, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 48, y: 192, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 64, y: 192, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 80, y: 192, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 96, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 192, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 208, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 224, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 224, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 224, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 224, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 224, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 224, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 224, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 224, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 224, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 224, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 224, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 224, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 224, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 208, y: 224, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 224, y: 224, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 240, y: 224, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 256, y: 224, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 272, y: 224, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 288, y: 224, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 304, y: 224, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 320, y: 224, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 336, y: 224, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 352, y: 224, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 368, y: 224, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 0, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 16, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 32, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 48, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 64, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 80, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 96, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 112, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 128, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 144, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 160, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 176, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 192, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 208, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 224, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 240, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 256, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 272, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 288, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 304, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 320, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 336, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 352, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 368, y: 240, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 0, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 256, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 272, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 272, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 336, y: 272, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 352, y: 272, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 368, y: 272, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 0, y: 288, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 16, y: 288, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 32, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 288, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 304, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 320, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 320, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 320, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 320, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 320, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 320, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 320, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 320, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 144, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 160, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 176, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 192, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 208, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 224, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 240, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 256, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 272, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 288, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 304, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 320, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 336, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 352, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 368, y: 320, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 0, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 16, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 32, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 48, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 64, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 80, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 96, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 112, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 128, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 144, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 160, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 176, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 192, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 208, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 224, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 240, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 256, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 272, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 288, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 304, y: 336, quality: high, premul: undefined, flip: true
CONSOLE MESSAGE: line 228: x: 320, y: 336, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 336, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 336, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 336, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 128, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 144, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 160, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 176, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 192, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 208, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 224, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 240, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 256, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 272, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 288, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 304, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 320, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 336, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 352, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 368, y: 352, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 0, y: 368, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 16, y: 368, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 32, y: 368, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 48, y: 368, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 64, y: 368, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 80, y: 368, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 96, y: 368, quality: high, premul: undefined, flip: false
CONSOLE MESSAGE: line 228: x: 112, y: 368, quality: high, premul: undefined, flip: false
